mychat/requirements.txt
agora-token-builder==1.0.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
python-dotenv==1.0.0
sqlparse==0.4.4
typing-extensions==4.8.0
